Refactor (5)
