+++ Source/Testably.Expectations/Core/ExpectationBuilder.cs
@@ -3,6 +3,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Testably.Expectations.Core.Constraints;
 using Testably.Expectations.Core.EvaluationContext;
+using Testably.Expectations.Core.Helpers;
 using Testably.Expectations.Core.Nodes;
 using Testably.Expectations.Core.Sources;
