--- Source/Testably.Expectations/Core/GrammaticVoice.cs
@@ -1,17 +1,0 @@
-﻿namespace Testably.Expectations.Core;
-
-/// <summary>
-///     Mode for English Grammar
-/// </summary>
-internal enum GrammaticVoice
-{
-	/// <summary>
-	///     The text should be in active voice.
-	/// </summary>
-	ActiveVoice,
-
-	/// <summary>
-	///     The text should be in passive voice.
-	/// </summary>
-	PassiveVoice
-}
