+++ Source/Testably.Expectations/Core/Helpers/BecauseReason.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Testably.Expectations.Core.Constraints;
 
-namespace Testably.Expectations.Core;
+namespace Testably.Expectations.Core.Helpers;
 
 internal struct BecauseReason(string reason)
 {
