+++ Source/Testably.Expectations/Core/Nodes/AndNode.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Testably.Expectations.Core.Constraints;
 using Testably.Expectations.Core.EvaluationContext;
+using Testably.Expectations.Core.Helpers;
 
 namespace Testably.Expectations.Core.Nodes;
