+++ Source/Testably.Expectations/Core/Nodes/Node.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using Testably.Expectations.Core.Constraints;
 using Testably.Expectations.Core.EvaluationContext;
+using Testably.Expectations.Core.Helpers;
 using Testably.Expectations.Core.Sources;
 
 namespace Testably.Expectations.Core.Nodes;
