+++ Source/Testably.Expectations/That/Delegates/ThatDelegateShould.NotThrow.cs
@@ -2,6 +2,7 @@
 using Testably.Expectations.Core;
 using Testably.Expectations.Core.Constraints;
 using Testably.Expectations.Core.Helpers;
+using Testably.Expectations.Core.Sources;
 using Testably.Expectations.Results;
 
 // ReSharper disable once CheckNamespace
